+++ reweave/tools/build-images.sh
@@ -58,7 +58,8 @@ build_image() {
 
     # shellcheck disable=SC2086
     docker buildx build \
-            ${POSTBUILD} \
+						${POSTBUILD} \
+						--sbom=true --provenance=true \
             --progress=plain \
             --target="$1" \
             --build-arg=ALPINE_BASEIMAGE=${ALPINE_BASEIMAGE} \
@@ -75,10 +76,12 @@ build_image() {
 
 # shellcheck disable=SC2086
 {
+docker buildx create --use --name buildkit
 build_image "weaverimage" ${WEAVER_IMAGE}
 build_image "weavexecimage" ${WEAVEEXEC_IMAGE}
 build_image "weavekubeimage" ${WEAVEKUBE_IMAGE}
 build_image "weavenpcimage" ${WEAVENPC_IMAGE}
 build_image "weavedbimage" ${WEAVEDB_IMAGE}
 build_image "networktesterimage" ${NETWORKTESTER_IMAGE}
+docker buildx rm buildkit
 }